requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv>=1.0.0

# Optional dependencies for Excel export (install if needed)
# pandas>=2.0.0,<2.1.0
# openpyxl>=3.0.0

# Optional dependencies for advanced browser automation (install if needed)
# playwright>=1.40.0
# After installing: playwright install chromium
